# eth/tracers/js: goja tracer