Latest bench:

```
name                                old time/op    new time/op    delta
Tracers/create-8                      9.34ms ± 3%    5.16ms ± 1%    -44.82%  (p=0.008 n=5+5)
Tracers/deepCalls-8                   63.1ms ± 1%    38.9ms ± 0%    -38.39%  (p=0.008 n=5+5)
Tracers/delegatecall-8                13.4ms ± 1%     7.7ms ± 1%    -42.80%  (p=0.008 n=5+5)
Tracers/innerCreateOogOuterThrow-8    11.8ms ± 0%     6.2ms ± 1%    -47.04%  (p=0.016 n=5+4)
Tracers/innerInstafail-8              9.35ms ± 8%    5.92ms ± 1%    -36.66%  (p=0.008 n=5+5)
Tracers/innerThrowOuterRevert-8       13.5ms ± 1%     7.2ms ± 2%    -47.16%  (p=0.008 n=5+5)
Tracers/oog-8                         10.4ms ± 1%     5.4ms ± 1%    -48.27%  (p=0.008 n=5+5)
Tracers/revert-8                      8.19ms ± 2%    4.88ms ± 0%    -40.39%  (p=0.008 n=5+5)
Tracers/revertReason-8                9.33ms ± 1%    5.16ms ± 0%    -44.72%  (p=0.008 n=5+5)
Tracers/selfdestruct-8                7.76ms ± 1%    5.21ms ± 1%    -32.91%  (p=0.016 n=5+4)
Tracers/simple-8                      11.0ms ± 6%     6.5ms ± 1%    -40.94%  (p=0.008 n=5+5)
Tracers/throw-8                       9.37ms ± 6%    5.16ms ± 0%    -44.93%  (p=0.008 n=5+5)

name                                old alloc/op   new alloc/op   delta
Tracers/create-8                       134kB ± 1%    2464kB ± 0%  +1738.41%  (p=0.008 n=5+5)
Tracers/deepCalls-8                   2.14MB ± 0%   10.61MB ± 0%   +396.56%  (p=0.008 n=5+5)
Tracers/delegatecall-8                 250kB ± 1%    3027kB ± 0%  +1111.10%  (p=0.008 n=5+5)
Tracers/innerCreateOogOuterThrow-8     333kB ± 0%    2701kB ± 0%   +711.24%  (p=0.008 n=5+5)
Tracers/innerInstafail-8              69.1kB ± 0%  2713.1kB ± 0%  +3828.84%  (p=0.008 n=5+5)
Tracers/innerThrowOuterRevert-8        284kB ± 1%    2831kB ± 0%   +896.80%  (p=0.008 n=5+5)
Tracers/oog-8                          154kB ± 1%    2437kB ± 0%  +1479.09%  (p=0.008 n=5+5)
Tracers/revert-8                      57.7kB ± 1%  2393.7kB ± 0%  +4044.88%  (p=0.008 n=5+5)
Tracers/revertReason-8                 109kB ± 2%    2420kB ± 0%  +2118.34%  (p=0.008 n=5+5)
Tracers/selfdestruct-8                39.2kB ± 1%  2527.6kB ± 0%  +6352.31%  (p=0.008 n=5+5)
Tracers/simple-8                       156kB ± 1%    2738kB ± 0%  +1658.64%  (p=0.008 n=5+5)
Tracers/throw-8                       95.8kB ± 1%  2416.2kB ± 0%  +2423.34%  (p=0.008 n=5+5)
```