@fjl 2021-10-19
Very interesting. I think some of the allocs are because of ToValue, but we need to measure it.