@fjl 2021-10-21
This looks very promising. I also looked into this a tiny bit today and found that another cause of slowdown is with BigInteger.js. Since goja is a bit slower than duktape at running actual JS computation, bigint operations take more time.

We could kind of work around this by reimplementing bigInteger.js with math/big, but not sure if it would really save much time/allocs in the end.