@holiman 2021-11-03
One thing that would be good to benchmark is: 

- Instantiation of a tracer, with (and without) preloading the `bigIntegerJS`. In duktape, that happens within `tracer.New`:

```
	tracer.vm.EvalString(bigIntegerJS)
	tracer.vm.PutGlobalString("bigInt")
```
On my machine, doing `tracer.New` takes `6-7 ms` _with_ bigint_ , but `0.8ms` without loading that. So it's a very non-negligible part of the time spent. I'm curious if it has the same (or larger) toll on goja. 

If it turns out that the actual execution of bigint js is a large part of the time, then other routes to investigate could be:
- Is it possible to evaluate only once, and somehow lock it in a global env and reuse that env? 
- Is it possible to clear and reuse one tracer instance (or perhaps a small pool of them)?

I suspect both of these options are easier with Goja, since we have more control over the code executing, being go and all. 

Also, https://github.com/olebedev/go-duktape is now archived and abandoned, which I think is fairly recent, which adds support for the case of switching away from it.