@holiman 2021-11-03
```
$ benchstat duktape-bigint goja-bigint 
name         old time/op  new time/op  delta
JSVM/all-6   7.00ms ±11%  4.00ms ± 2%  -42.86%  (p=0.000 n=10+9)
```
So on my machine, tracer instantiation (plus bigint preloading) goes from `7ms` to `4ms`.