@holiman 2021-11-03
Oh wow, bigint js really takes a toll on goja too: 
```
$ benchstat goja-bigint goja-nobigint 
name        old time/op  new time/op  delta
JSVM/all-6  4.00ms ± 2%  0.21ms ± 0%  -94.63%  (p=0.003 n=9+4)
```