@s1na 2021-12-09
So I enabled the JS tracer [tests](https://github.com/ethereum/go-ethereum/blob/ae8ff2661d917506ec29bd8f457884d483e2ce42/eth/tracers/js/tracer_test.go) for the goja tracer. These 2 cases fail:

```
=== RUN   TestGojaTracer                                                                                                                                                                                                                 tracer_test.go:135: testcase 0: expected return value to be '[{},{},{}]' got '[[],[],[]]', error to be '' got ''                                                                                                                                 code: {depths: [], step: function(log) { this.depths.push(log.memory.slice(-1,-2)); }, fault: function() {}, result: function() { return this.depths; }}                                                                 tracer_test.go:135: testcase 7: expected return value to be '' got '{"foo":{"foo":...}}}', error to be 'RangeError: json encode recursion limit    in server-side tracer function 'result'' got ''
                code: {step: function() {}, fault: function() {}, result: function() { var o={}; var x=o; for (var i=0; i<1000; i++){  o.foo={}; o=o.foo; } return x; }}
```

The second one is because of the limitations in duktape, so not really a failure. The first one is a bit weird because I expected a slice of memory to be a list `[]`, not an object `{}`. Then I checked and it seems in duktape in most cases where we push a buffer from go to js (i.e. memory, any address or hash with fixed length like `contract.GetCaller()`) it shows up on the js side as something like `{0: 1, 1: 7, 2: 5...}`, whereas the current goja implementation simply returns sth like `[1, 7, 5...]`. I find the current behaviour a bit weird, anyone knows why that's the case?