@s1na 2021-12-09
I've been tracing goerli transactions via both duktape and goja to compare the result for our various tracers. So far found a difference in prestate tracer (will update this list):

- https://gist.github.com/s1na/7c9bd636c3ac0722a26a81a281d335d8