@s1na 2022-04-07
I fixed one of the main differences we had between goja and duktape (see [comment](https://github.com/ethereum/go-ethereum/pull/23773#issuecomment-989903232)). Basically byte strings were being passed to JS as normal arrays and not buffers (or more concretely `Uint8Array`). I found no way to create these objects using Goja's go API https://github.com/dop251/goja/issues/379) so I'm doing a hack for now. And that's we have a JS function that takes in an array and returns a `Uint8Array` from it. We call that JS func `toBuf` from Go whenever we want to push a buffer to JS.