@s1na 2022-04-13
Note for reviewers: you'll see `panic`s dispersed in the code. They're all in native functions that are exposed to JS. Goja will raise those panics as a JS interrupt which can be caught on the JS side. If they're not they'll simply cause execution to stop and the error will be returned like any other JS error.