@s1na 2022-05-02
So for some reason the `prestateTracer` is slower in goja than in duktape. No idea why yet, if I had to guess it's inefficiencies in how built-in functions like `toContract`, etc. are implemented. For every other tracer goja is a bit faster:

```
# Duktape

       noopTracerLegacy  callTracerLegacy  prestateTracerLegacy  opcountTracer  unigramTracer
mean          68.165723         98.223947             96.123697      68.585523      79.273222
std           39.852465         86.367623             74.865044      40.145431      53.258374
min            8.000000          9.000000              9.000000       8.000000       8.000000
25%           36.000000         44.000000             46.750000      36.000000      40.000000
50%           61.000000         79.000000             81.000000      61.000000      70.000000
75%           93.000000        129.000000            126.000000      93.000000     107.000000
max          376.000000       1918.000000           1709.000000     423.000000     963.000000

# Goja

       noopTracerLegacy  callTracerLegacy  prestateTracerLegacy  opcountTracer  unigramTracer
mean          60.582691         77.329746            154.703670      61.275940      66.597191
std           37.869345         54.091886            146.952691      37.834356      40.654601
min            5.000000          6.000000              7.000000       5.000000       5.000000
25%           30.000000         38.000000             61.000000      30.000000      34.000000
50%           53.000000         68.000000            108.000000      54.000000      59.000000
75%           85.000000        105.000000            212.000000      86.000000      92.000000
max          256.000000       1366.000000           3543.000000     206.000000     446.000000
```